crash: Experiment with using daemon-store.
In order to ensure all crashes that occur while a user is logged in are
encrypted, put them in the daemon-store.

This also has the benefit of allowing crash_reporter and crash_sender to
read/write crashes without needing access to the user's home directory,
which is necessary for de-rooting the crash system.

Only put half of crashes in the daemon-store, so that we can compare
those that do and don't use daemon-store and make sure that the
daemon-store solution works.

Also, minor refactor to simplify vm_collector's logic around when to use
daemon_store.

BUG=b:186659673
TEST=tast run forward_crbook arc.AppCrash* arc.SELinuxViolation \
          crash.* ui.ChromeCrash* ui.ChromeCrash* ui.WebUIJSErrors
TEST=tast run $eve_crbook wifi.CheckIntelFWDump

Cq-Depend: chromium:2333434
Change-Id: I2d136eb4278f3e296b2a42d7516ab89bd501fd19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2333354
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>