ml: Add multiprocess support
The overall pattern is control-worker: there is one control
process (the original ml-service process) that spawns and manages the
worker processes. For more details, please see
go/cros-ml-multiprocess.

Currently, each (and only each) WebPlatformHandwritingModel is run in
a dedicated process. The worker process will be created when the
`LoadWebPlatformHandwritingModel` API is called. Then a mojo
connection between the browser and the worker process is
constructed. The lifetime of the worker process is bound to this mojo
connection, that is, the worker process will exit in the
disconnection handler of this mojo connection.

And we may roll this to other models in the following CLs.

BUG=chromium:933017
TEST=on device (nocturne and kukui), multiprocess works for
TEST=WebPlatformHWR model.

Cq-Depend: 2516101
Change-Id: Ib517752dc69af60f0ed12cbeeb9be6c955517990
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2854267
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>