shill: add always-on VPN configuration to Profile.
The always-on VPN configuration is represented by two properties (the
mode and the service) in the Profile. It exposes them in its D-Bus
interface as a string (the mode) and an object path (the service). When
one of the properties changes, the manager is notified the profile
changed.

As the configuration is user related and not service specific, it is
persisted in the profile storage under the "global" entry name. The mode
is stored as is, while we only keep the storage identifier for the
service. Both properties are persisted when the profile itself is saved.

BUG=b:167661214
TEST=Use dbus-send to trigger SetProperty/GetProperties methods on
  org.chromium.flimflam.Profile to modify and check the two properties on
  /profile/chronos/shill.  Check modified properties are still have a
  non-default value after a reboot (persistence works).
  FEATURES="test" emerge-${BOARD} shill

Change-Id: I55285f2523411a43134aff298caaac92e20594de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2565910
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>