biod: Refactor BiodStorage record reading
As mentioned in the removed TODO, it's a bit awkward to use a callback
to synchronously load records. The code is easier to follow if we return
the records directly when we do the read.

As part of the change, this also refactors
CrosFpBiometricsManager::InternalRecord to BiodStorage::RecordMetadata,
since the object is a low-level representation of what is stored on disk
and closely tied to the BiodStorage code.

We use return values instead of out parameters since that is consistent
with the style guide and recommended in https://abseil.io/tips/176.

BUG=b:168340872
TEST=cros_workon_make --board hatch biod --test

Change-Id: Ide59609d020227c581121c625ae16159ebd6b819
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2373302
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Yicheng Li <yichengli@chromium.org>