dns-proxy: Introduce dns-proxy service
This patch adds the initial framing of the dns-proxy service including
the process lifecycle management for the main controller and child proxy
processes, and launches and keeps running the system proxy.

Note that until an rdep is introduced, this package will not be included
in the system image.

BUG=b:174802586
BUG=b:171157837
TEST=with supporting patches flashed hatch, verified manaually

Cq-Depend:chromium:2583585
Change-Id: I43f36d207b5a4a6e50e5e7488ce033e18ab92389
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2581753
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>