camera: camera3_test: Add TakePictureZslTest
Adds a test to verify that pictures can be taken successfully with ZSL
turned on. We take a series of images and set an interval of 500ms to
give ZSL room to accumulate (attach) buffers. The test is skipped if ZSL
is not supported.

BUG=b:151048328
TEST=tast run $DUT camera.HAL3StillCapture*'

Change-Id: I8b241a4a7759acee6a266cb95c9bdf43108c6a9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2587120
Tested-by: Jasmine Chen <lnishan@google.com>
Commit-Queue: Jasmine Chen <lnishan@google.com>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Auto-Submit: Jasmine Chen <lnishan@google.com>