cryptohome: Stop accessing flags directly in SerializedVaultKeyset.
This removes accesses to the flags directly in the |serialized_|
member of the VaultKeyset class. This is a step towards removing all
calls to serialized(), and thus direct access to |serialized_| entirely.
This CL was previously reverted and fixed to reset LE credentials
properly.

BUG=chromium:1151512
TEST=tast test ${DUT_IP} hwsec.*

Change-Id: Icf99136bcb50d57b9627fe193c1a1f288e3a4110
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2721059
Tested-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>