cryptohome: Log UMA histogram for LE log entry replay count
We'd like to make sure that we have sufficient margin in number of log
entries we keep to recover from all but the most egregious pinweaver
state sync loss events occurring in the field. At a histogram to
collect data on how many entries we need to replay.

BUG=b:176465711
TEST=Force a log replay and check chrome://histograms

Change-Id: Ie5e46a26088a4c06b4e62c2c4e9f86c09d06fff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2611743
Commit-Queue: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>