biod: Change encrypted fp templates to std::vector
The ELAN515 sensor's template size is larger than 65536 bytes, and
allocating a brillo::SecureVector of this size will crash biod.
Changing RLIMIT_MEMLOCK is also not an option because the current
limit is already the hard limit set by kernel.

Templates are encrypted by FPMCU so it's not strictly necessary to
use SecureVector for templates in biod.

BUG=b:175158241
TEST=none

Change-Id: Ib28d0c7955c5f78db3bee63bc0f4cc55c320c4db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2627733
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>