iioservice: Assume requested frequency as EC device frequency
EC returns the frequency the sensor is running at:

- it can be higher for internal needs (EC set BMI160 frequency at 400Hz
for double tap detection).
- it can be different if the EC is delayed to set the frequency (frequency
is set in the motion task, not the host task).

Therefore, use the requested frequency as based for calculation.
The sample rate can be higher than expected, but the applications should
handle it: CTS pass criteria is 90%-220% of requested frequency.
[For instance, if one client set ODR at 10Hz, and the other at 150Hz],
iioservice will send every 15th sample (averaged) to the slow client.
But EC would set the ODR at 200Hz, so the slower client real ODR will be
13.3Hz.]

BUG=b:172208566, b:177707650
TEST=unit tests and run on kohaku

Change-Id: Ia91b46e8f5cb1f3146dbb32e7c1ea5241a30c737
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2616048
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>