shill: Add support to reset QCom qcom-q6v5-mss based modems using sysfs
Qcom q6v5 modems cannot be reset using dms commands,
the only way to reset them is to use sysfs. When shill
tries to reset a cellular device, we check to see if the
sysfs corresponding to q6v5 modem is present.
If it is present then we use the state file to reset the modem.

BUG=b:168113309
TEST=emerge-trogdor shill && cros deploy

Change-Id: Ife2f8ad445ec5e590ebeb1b22e4c2b485e88f9db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2620422
Commit-Queue: Madhav . <madhavadas@google.com>
Tested-by: Madhav . <madhavadas@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Madhav . <madhavadas@google.com>