shill: delay always-on VPN connection on failure.
If the service configured as always-on VPN fails to connect, it may end
in a connect->failure loop. To avoid it, the failure was considered as a
black hole: when the client failed, it was not reconnecting.

This changeset implements a retry mechanism, with a backoff time between
retries, to prevent keeping the device (or the server) busy with failing
attempts. The number of attempts is counted and used to compute a
backoff time between retries, up to ~60s. When a failure occurs, the
service is not immediately reconnected, but a connect task is scheduled.
A connection success, a physical network change or a configuration
change will reset the backoff/retry mechanism.

BUG=b:167661214
TEST=Configure an OpenVPN client as the always-on VPN. The device should
be able to reach internet through ethernet and wifi. Shutdown the server
(to trigger future failures) and unplug ethernet. Ensure connection
attempts occur, delayed by an increasing time (using the logs). Restart
the server, check the client connects successfully on the next attempt.
Plug the ethernet back, check the client reconnects immediately.

Change-Id: I86f0300d3c663ecf64ba3fdc3630e22495bc8821
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2624887
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>