verity: Merge dm-bht-userspace.h to dm-bht.h
The three functions defined in dm-bht-userspace do not seem to be really
only for user space. Looking at the platform2/installer use case for
verity, it is obvious that other functions in dm-bht.h are used as
well. So it's better to just have all functions in one file so we can
properly convert to a better C++ style (with classes and everything).

BUG=chromium:886953
TEST=cros_workon_make --board reef --test verity

Change-Id: Ic9c3b187f857c46ea69cb2866b69f2c4ee2701c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2627328
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>