camera: Remove SHM buffer allocation from CameraBufferManager
We no longer need to allocate buffers backed by SHM. This CL removes the
code that allocates SHM buffers and the "type" field in
camera_buffer_handle.
Note that this is not ABI-compatible, and anything that depends on
libcbm would need to be rebuilt.

BUG=b:177534317
TEST=FEATURES=test emerge-eve cros-camera-libcbm

Cq-Depend: chromium:2629196
Change-Id: I0743922ddac5d315aad08079e7c10b72e5db6376
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2629192
Commit-Queue: Jasmine Chen <lnishan@google.com>
Tested-by: Jasmine Chen <lnishan@google.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>