cryptohome: Call selective drop cache on the stateful partition mount
Since CL:2600839, instead of using the shadow root directory, we call
InvalidateDirCryptoKey() on the encrypted directory instead of the
ShadowRootDir(). This is the correct usage: in absence of the
filesystem encryption key reference, the directory is used to
fetch it.

However, when the encrypted directory is used to call
FS_IOC_DROP_CACHE, it fails to clear the pagecache/prune immediate
child dentries since the encrypted directory has an open file
descriptor. This results in the child dentries not getting pruned
on the above ioctl call and appearing as plaintext. Instead, always
call DropMountCaches() on the stateful partition mount point.

BUG=b:177708224
TEST=login_CryptohomeDataLeak passes

Change-Id: Ifc985fdc9a15fbbb048dea2f7a3b6cc09ddd1a10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2637734
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>