typecd: Make mode entry code aware of session state
Update the mode entry code to be aware of whether a user is active or
not. Based on that we can decide whether to enter TBT or DP mode (based
on what modes are available for the peripheral).

This should not affect current one-shot mode entry, since the
|user_active_| field is fixed at "true" throughout daemon life-cycle,
and not changed.

This will allow us to write tests for the various other scenarios
(lock/unlock/logout/login) without affecting production code. We will
develop these as separate callbacks and only later connect them to the
session manager callbacks.

BUG=b:171839508
TEST=Unit tests should continue to pass.

Change-Id: Ie6d1728a8ac9f1b5c17b6a266c2bc951338885d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2639213
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>