power: Always enable proximity events if possible
There are some IIO proximity devices that need to be enabled but we
currently do that via udev rules. Let's always enable the threshold
event as that's what this code is watching for. Existing udev rules will
still work and this will either write a 1 to enable it again or fail and
stick a warning into the logs. In the future we can update the cros
configs for boards that use udev rules to set the proper channel so the
right event can be enabled.

Note that sometimes we really do need cros config to tell us which
channel to enable threshold events for, because there are multiple
channels for the IIO device. If that's the case we stick the warning in
the logs so we can migrate those boards to using the cros config to tell
us which channel to enable.

BUG=b:173721158
TEST=emerge-strongbad power_manager && cros deploy; See event enabled in
     sysfs

Change-Id: I21622707239947fcbaadeafc33145b8f0cbe6f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2644385
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>