shill: avoid transition back to the Connected state in SetupConnection
The Connected state is less informative than any other Connected state,
so there's no reason to transition back to it when we are already in a
Connected state. Instead of disturbing the state machine, and
potentially the Service sort order, leave it as is until portal
detection finishes and transitions the Service to either the Online or a
portalled state.

BUG=b:177500540
TEST=unit tests
TEST=wifi_matfunc
TEST=wifi.RoamFT no longer repros the issue in $BUG

Change-Id: I86871a2dd03b4d2f6ea9597cfbb18737ea42a08b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2644498
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>