shill: Add dns-proxy support to Resolver.
Adds a method to Resolver for the Manager to toggle use of the dns-proxy
address(es) as the name servers for the system resolver. This will
subsequently be wired to the relevant Manager property(ies).

As elaborated in the design doc, this method is preferred so the Manager
can control the global setting independently of the Connection data
model.

BUG=b:171362408
TEST=units

Change-Id: I53d1e5e6e360b30863a402a469209e13497da28f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2644554
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>