shill: vpn: Move interface management back to drivers
This patch moves the VPN interface creation and deletion from VPNService
back to VPNDrivers, since different drivers actually differ a lot in it
(e.g., how and when the interface is set up). Note that the management
of VirtualDevice is still in the VPNService.

See the new comment for VPNDriver for more details.

BUG=b:178766289
BUG=b:177876632
TEST=Manullay verified VPN connections for all the four types on a DUT.

Change-Id: I2a1cfc00e6cdb9711b196b3e7a76b01b8b8e5b16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2646785
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>