shill: Set/Clear dns-proxy address
This patch adds logic to Manager to propagate the dns-proxy address
property changes to the resolver. It also updates the same whenever the
default physical service goes online or offline. This is necessary to
prevent situations like captive portal negotiation from using the proxy.

This change also introduces the mocked resolver into the ManagerTest
since the path is actually setup in Start, but this isn't called in the
test (the running flag is manually flipped).

BUG=b:171362408
TEST=units

Change-Id: Ice3278095adec4aab337139ff6d8502a339b59cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2649366
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>