shill: Remove CHECK
This patch changes a CHECK in ClearDNS to LOG(DFATAL) and adds the
similar block in Emit (seems like it should have been there already).
It then adds a CHECK on the actual filepath in Start which is what the
original intention appears to have been.

BUG=b:171362408
TEST=units

Change-Id: Ie138403ed0ffb9d7085154a9d89ea454939ab58f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2649365
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>