arc: arcvm-forward-pstore copies pstore back to /run/arcvm
This change modifies arcvm-forward-pstore to copy the pstore file
back to /run/arcvm/arcvm.pstore before ARCVM restarts. This allows
the guest to generate /sys/fs/pstore/dmesg-ramoops-0 on next boot.

This change also modifies the lifecycle of this service: instead of
stopping on 'stopping arcvm-pre-login-services', it now persists across
ARCVM restarts in a single session. This is so that it can delete
/run/arcvm/arcvm.pstore on exit to prevent it from leaking out of the
user session.

BUG=b:177626889
BUG=b:177682644
TEST=tast run <DUT> arc.KernelCrash.mock_consent

Change-Id: Id6dbe007125a216d4e54cabe87690cc88f185cba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2648254
Tested-by: Willie Koomson <wvk@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Kimiyuki Onaka <kimiyuki@google.com>
Commit-Queue: Willie Koomson <wvk@google.com>