dns-proxy: Add curl client
This patch adds asynchronous DoH resolver proxy using
libcurl. DoHCurlClient receives a normal DNS query and
run it to the DoH server.

In this patch a query will be perform sequentially
on each DoH servers (runs asynchronously in the
background). When a query is done, a callback will
be called to the caller.

BUG=b:177375227
TEST=compiled

Cq-Depend: chromium:2651690
Change-Id: If188e943f15e57601a64b388dec4edef5441bfd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2651728
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>