dns-proxy: Add ares client
This patch adds a DNS proxy resolver that takes a
DNS query data and proxies it to DNS servers.

The resolve call is asynchronous and a callback will
be called upon completion.

If given multiple DNS servers, the current implementation
will not try to send query to all servers. It instead
will try to query each server until it finds a good one.

BUG=b:177375227
TEST=compiled

Change-Id: If08fcb43dfc5ff0c2b1db8b6a0362b38bbcd001c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2651729
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>