dns-proxy: Set shill's dns-proxy address property
When the system proxy is ready to start serving, it will set the
DNSProxyIPv4Property on shill. Likewise, when it shuts down it will
clear it. The controller also monitors the life time of the system proxy
to ensure the property is cleared whenever it (the system proxy) exits
under any circumstance.

Unit tests upcoming in next patch(es).

BUG=b:171157837
TEST=manual

Change-Id: I74200d023b4a495239919f735cf8e1fd65dc0ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2653747
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>