vm_tools/sommelier: add initial virtgpu channel implementation
This is the initial implemententation the virtgpu {wayland,
cross_domain} context type.  It is more or less based on the
trail-blazing virtwl device.

The key features of this approach are:

* Receives responses via a shared ring buffer of type
  BLOB_MEM_GUEST.

* Uses the upcoming virtgpu context type feature to route the
  command stream.

* Uses the upcoming virtgpu multiple timelines features for
  synchronization.

The uapi changes are just a prototype at this point, and it poses
little technical debt risk since the the virtgpu channel will not
be enabled until the changes are accepted into Linus' tree.  Thus,
this is just for fast iteration + testing.

BUG=b:150239451
TEST=Sommelier testing protocol on Borealis:

https://goto.google.com/sommelier-testing

Change-Id: Ic1aae65855ccca2d54d63321607a1e878b94b7be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2669422
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>