shill: cellular: ignore MTU override if network reports lower MTU
Right now, if there is an MTU override listed in the APN database,
we will always use it even if the network reports a lower MTU. This
leads to issues with fragmentation and in some cases certification.
We will still use the override if the network reports a higher MTU
or none at all.

Also, symbolize the places where 0 is used as an MTU value by
replacing them with IPConfig::kUndefinedMTU for clarity.

BUG=b:177039849
TEST=unit tests, manual inspection, cellular_Smoke.att in the lab

Change-Id: I98853fd91aaaf8bcd2fae10759ea6d603eb395a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2657052
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>