vm_tools/sommelier: remove SHM_DRIVER_DMABUF
If we are assuming sommelier only operates in a virtualized
environment, SHM_DRIVER_DMABUF is unlikely to work.  The reason
is because this driver creates it's buffers using gbm, which
when virtualized, translates to virtgpu.  virtgpu buffers created
by libgbm can not be mapped directly (though it may work for
buffers created by libminigbm, but that is not used by Crostini).
As such, this is dead code.

BUG=b:150239451
TEST=Sommelier testing protocol on Borealis:

https://goto.google.com/sommelier-testing

Change-Id: Ica5a0d8f4253854acfe21c7fddbf5d4925972c12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2669424
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Commit-Queue: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>