dns-proxy: Add more logic around setting shill property
This patch adds handling for when shill resets/crashes and attempts to
(re)set the address property. It also changes the setting logic from a
single call to a delayed retry... since we really need this value
correct. On exceeding the max number of retries, we will simply crash
the process when setting up the system proxy for the first time. This
will restart everything and hopefully subsequently succeed.

BUG=b:171157837
TEST=manual

Change-Id: I7880b29d753885ea9a1b1f1015d0d50f1ddf7257
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2666548
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>