dns-proxy: Run CURL resolve concurrently for each DoH server
When multiple DoH servers are given, use all of the servers
concurrently to resolve the query. This allow DNS resolving
to run correctly even if a bad server is provided.

Upon getting the result of a query, DNS proxy will proxies
the result if:
- the query is first successful query OR
- the query is the last query and no successful query was
  done previously.

BUG=b:177375227
TEST=dig <URL> @<DNS_PROXY_ADDR>; watch tcpdump;

Change-Id: I1bb718f5e7acea7215366ab363f4c1bb693e27d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2661139
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>