dns-proxy: Run ares resolve concurrently for each DNS server
When multiple DNS servers are given, use all of the servers
concurrently to resolve the query. This allow DNS resolving
to run correctly even if a bad server is provided.

Upon getting the result of a query, DNS proxy will proxies
the result if:
- the query is first successful query OR
- the query is the last query and no successful query was
  done previously.

This patch also adds timeout for ares resolving.

BUG=b:177375227
TEST=dig <URL> @<DNS_PROXY_ADDR>; watch tcpdump;

Change-Id: I2d59fda085ef159ece828191cfb6f7b4a3448088
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2661140
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>