shill: Change signature of EventDispatcher::Post*Task()
This patch changes the second parameter of EventDispatcher::Post*Task()
from base::Closure to base::OnceClosure, to match the signature of
base::TaskRunner::Post*Task().

There is an implicit conversion from base::(Repeating)Closure to
base::OnceClosure, but not vice versa, so this patch has no side effect
on the current callers of this function, and allows us to post
OnceClosure in the future code.

BUG=b:178766289
TEST=unit_tests

Change-Id: I3d87f9a6593e1526bc75e10ab2fc3f220ed54593
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2662869
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>