shill: cellular: remove empty APN from the list
Some modems might return an APN with an empty name in the MBIM
ProvisionedContext. In the LTE modem, this empty APN is the default
attach APN for most networks.
On the shill side, we don't really need in the list and it confuses the
Chrome code when sent over D-Bus in the Cellular.APNList (see bug).
So just skip it.

BUG=b:176226595
TEST=unit-tests (the new one fails without the change, passes with it)
TEST=manual, with NL688 no longer see a crash when switching APN in the
UI.

Change-Id: I4adff3d80a35de8ad561e242b4b60146816cd4a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2663196
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>