power: Claim ownership of config in user_proximity_watcher
The config_ pointer goes out of scope and gets destroyed, but this class
doesn't know that. Once we try to use it we get a crash. Move the
ownership to the proximity watcher class so that it can keep using the
config later on, like when a udev event triggers indicating that a
sensor has been detected.

BUG=b:166572529
TEST=emerge-trogdor power_manager; cros deploy $DUT power_manager; check
     /var/spool/crash/ for powerd crashes and see none

Change-Id: Ib340d44e86519e7dbe5c3072800cb67901e941cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2674052
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>