vm_tools: Only connect to arc-powerctl after VM upgrade
This change modifies the ARCVM controller in Concierge so that it only
attempts to connect to arc-powerctl after the VM has been upgraded
(SetVmId has been called for the VM). If the VM has not been upgraded
yet, it skips to un-graceful shutdown (crosvm socket, TERM and KILL
signals).

This change also adds support for retrying connecting to arc-powerctl,
to cover the case in which the VM has upgraded, but arc-powerctl is
still starting.

BUG=b:172266394
TEST=Flash modified arc-powerctl (ag/13473062). Log in, then "stop ui".
 Check logs to make sure concierge retries connecting to arc-powerctl.

Change-Id: I14bb603cb51dea45797904613e928d061680bfcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2674423
Tested-by: Willie Koomson <wvk@google.com>
Commit-Queue: Willie Koomson <wvk@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>