shill: vpn: Extract connect timeout logic into VPNService
The connect timeout logic is the same for all the VPN drivers:
- A timeout should be started after the driver enters connecting or
  reconnecting state;
- A timeout should be stopped after the driver leaves connecting or
  reconnecting state (disconnect, connection success or failure).

This patch moves the code for timeout in the base VPNDriver class and
the code for triggering timeout in the child classes into VPNService.
Different drivers have different timeout values, that will be returned
by ConnectAysc() or OnDriverReconnecting(), and when the timeout is
triggered, VPNDriver::OnConnectTimeout() will be invoked.

BUG=b:178766289
TEST=unit_tests
TEST=manually verified that the VPN connection still works

Change-Id: I0444aeaa4bedeccdf2c3da265bb8da0ad399b2b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2675401
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>