arc: Use mount-passthrough-jailed-media to set up external storage files
In ARCVM, we need to set up the external storage files (MyFiles and
removable media) so that they are owned by Android's MediaProvider.
This CL introduces a new script mount-passthrough-jailed-media, which
obtains the MediaProvider UID by inspecting the MediaProvider package
directory, and passes the value to mount-passthrough. The script is run
by the Upstart jobs to set up the external storage files, namely
arcvm-mount-myfiles and arcvm-mount-removable-media.

Note that this is a short-term quick fix. The setup will be implemented
in a more robust way in the future.

BUG=b:171727981
TEST=Start ARCVM via (1) normal sign-in, (2) opt-out and opt-in,
TEST=(3) $ pkill crosvm, and (4) $ adb reboot. Confirm that the external
TEST=storage files are properly set up and shared with ARCVM.

Cq-Depend: chromium:2676236
Change-Id: Ib938757e6b163a0506d12f9718c6730a6a32c6fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2677866
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>