shill: Remove incorrect vpn-source routing rule
When per device-table was introduced in crrev/c/1590183/, a source-
matching routing rule was added with the intent of helping vpn
traffic choose underlying physical interface correctly.

However, this was never actual needed and also implemented wrong.
This rule never matches anything and the source ip of encapsulated
vpn traffic is chose correctly due to the following catch-all rule.

This patch remove this rule. If catch-all rule is to be removed,
fwmark should be the correct way to match encapsulated vpn traffic.

BUG=b:174661113
TEST=unit, verify routing w/split-tunnel and full tunnel vpn

Change-Id: I4da47ae5417efbbd0ebe4eddd44670e3fb7c9166
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2677869
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>