cryptohome: Replace GetProcessesesWithOpenFiles() with ExpireMount()
GetProcessesWithOpenFiles() function is called by mount_helper.cc,
service.cc and userdataauth.cc and make cryptohomed and
cryptohome_namespace_mounter processes interact with the whole Chrome
OS system. In turn this requires 400+ SELinux allow rules for
cros_cryptohomed and cros_cryptohome_namespace_mounter to be added
to their SELinux domains which are otherwise not needed. The reason
why we would need so many rules is because /proc/$$/fd, which is the
source for ReadLink()  is labeled differently for different processes.
And it's difficult to scale the allowlist to include every new SELinux
context added to the allowlist. So we need to remove
GetProcessesWithOpenFiles() to unblock the SELinux work for cryptohome.

To keep the cleanup behavior unchanged, unmount() system function is
used with MNT_EXPIRE flag which doesn't unmount either active or
passive mount points in the first call but return a spesific error for
inactive mount points. This is used to determine the active mount
points to exclude them from the unmount list and from the pkcs11 token
removal list.

BUG=b:178041887
TEST=FEATURES=test emerge-DUT cryptohome
TEST=Set the flag 'mount_non_ephemeral_session_out_of_process_' to
false. Together with the dependent CL emerge and deploy cryptohome.
Sign in to DUT and crash cryptohomed with 'kill -SIGKILL' while
keeping a file in the user home directory open. Observe that
cryptohomed crash doesn't affect file access, i.e. read, write, save
and, the related mount points are not unmounted.

Cq-Depend: chromium:2670714
Change-Id: I2805821ed066dff4059f27fc4a0bc2cce078e9a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2654352
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>