biod: Add --seccomp arg to bio_crypto_init
This new option is only used when generating seccomp policies.
The purpose is to have bio_crypto_init exercise all code paths,
so that strace will see the full extent of syscalls used.

This is basically increasing "code coverage", so we may be able
to link this with unit tests at some point.

We must move the child process code into its own function
(with its own scope) in order to ensure that all deconstructors
are run in the child process, before we infinite loop.
Without this, we failed to see the munlock syscall on arm64.

BUG=b:178654555
TEST=Run run_bio_crypto_init_strace.sh from https://crrev.com/c/2686395

Change-Id: I3b3a35be934cbeddd48b1d63b55bc6189130fdc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2681679
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>