imageloader: Add a global context object
We can use this global context to get properties that we would need
througth the program so we don't have to pass objects/flags around too
much. The specific use case of this is finding out whether an image is
an official signed one or not.

BUG=b:174874518
TEST=cros_workon_make --board reef --test imageloader

Cq-Depend: chromium:2687886
Change-Id: If9897a1d9ae549b52dee05f044944e8e4f1ac271
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2688000
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>